+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ### 0.32 — Unreleased
 
+- The `force_open` option now defaults to true for working orders, and is automatically set to true when creating a
+  position that specifies a limit distance, limit level, stop distance, or stop level
 - Added `--to` option to the `ig_markets activities`, `ig_markets transactions`, and `ig_markets performance`
   commands
 - All timestamps provided to the `ig_markets activities`, `ig_markets transactions`, and `ig_markets performance`
   commands must now specify an explicit time zone
+- 
 - Support the `unavailable` state on `IGMarkets::Streaming::MarketUpdate#market_state`
 
 ### 0.31 — March 26, 2017